+++ dynapipe/estimatorCV.py
@@ -1,10 +1,11 @@
 #!/usr/bin/env python
 
 import pandas as pd
-from sklearn.linear_model import LogisticRegression,LinearRegression
-from sklearn.svm import SVC,SVR
+from sklearn.experimental import enable_hist_gradient_boosting
+from sklearn.linear_model import LogisticRegression,LinearRegression,HuberRegressor,RidgeCV,LassoCV,SGDRegressor
+from sklearn.svm import SVC,SVR,LinearSVR
 from sklearn.tree import DecisionTreeRegressor
-from sklearn.ensemble import RandomForestClassifier,RandomForestRegressor,AdaBoostRegressor,AdaBoostClassifier
+from sklearn.ensemble import RandomForestClassifier,RandomForestRegressor,AdaBoostRegressor,AdaBoostClassifier,HistGradientBoostingRegressor
 from sklearn.neural_network import MLPClassifier,MLPRegressor
 from sklearn.ensemble import GradientBoostingClassifier,GradientBoostingRegressor
 from sklearn.neighbors import KNeighborsRegressor
@@ -237,7 +238,8 @@ class reg_cv:
         #     'min_samples_leaf':[1,3,5]
         #     }
         parameters = para_data["reg"]["tree"]
-        return(GridSearchCV(tree_cv, parameters,cv=self.cv))       
+        return(GridSearchCV(tree_cv, parameters,cv=self.cv))
+      
     def ada(self):
         warnings.warn = warn
         ada_cv = AdaBoostRegressor()
@@ -261,4 +263,51 @@ class reg_cv:
         #     }
         parameters = para_data["reg"]["xgb"]
         return(GridSearchCV(xgb_cv, parameters,cv=self.cv))
+    # # New add August 5,2020
 
+    def hgboost(self):
+        warnings.warn = warn
+        hgboost_cv = HistGradientBoostingRegressor()
+        parameters = {
+            'max_depth': [3, 5, 7, 9],
+            'learning_rate': [0.1, 0.2,0.3,0.4]
+            }
+        # parameters = para_data["reg"]["tree"]
+        return(GridSearchCV(hgboost_cv, parameters,cv=self.cv)) 
+
+    def huber(self):
+        warnings.warn = warn
+        huber_cv = HuberRegressor()
+        parameters = {
+            'fit_intercept' : [True,False]
+            }
+        # parameters = para_data["reg"]["tree"]
+        return(GridSearchCV(huber_cv, parameters,cv=self.cv))
+
+    def rgcv(self):
+        warnings.warn = warn
+        rgcv_cv = RidgeCV()
+        parameters = {
+            'fit_intercept': [True,False]
+            }
+        # parameters = para_data["reg"]["tree"]
+        return(GridSearchCV(rgcv_cv, parameters,cv=self.cv))
+
+    def cvlasso(self):
+        warnings.warn = warn
+        cvlasso_cv = LassoCV()
+        parameters = {
+            'fit_intercept': [True,False]
+            }
+        # parameters = para_data["reg"]["tree"]
+        return(GridSearchCV(cvlasso_cv, parameters,cv=self.cv))
+    def sgd(self):
+        warnings.warn = warn
+        sgd_cv = SGDRegressor()
+        parameters = {
+            'shuffle': [True,False],
+            'penalty': ['l2', 'l1', 'elasticnet'],
+            'learning_rate': ['constant','optimal','invscaling']
+            }
+        # parameters = para_data["reg"]["tree"]
+        return(GridSearchCV(sgd_cv, parameters,cv=self.cv))